+++ src/os2datascanner/projects/admin/adminapp/models/scannerjobs/msgraph.py
@@ -269,6 +269,11 @@ class MSGraphSharepointScanner(MSGraphScanner):
         verbose_name=_("scan drives")
     )
 
+    scan_descriptions = models.BooleanField(
+        default=False,
+        verbose_name=_("scan list descriptions")
+    )
+
     supports_rule_preexec = True
 
     sharepoint_sites = models.ManyToManyField(
@@ -291,7 +296,8 @@ class MSGraphSharepointScanner(MSGraphScanner):
                 client_id=str(self.graph_grant.app_id),
                 tenant_id=str(self.graph_grant.tenant_id),
                 client_secret=self.graph_grant.client_secret,
-                sites=list(self.sharepoint_sites.values())
+                sites=list(self.sharepoint_sites.values()),
+                scan_descriptions=self.scan_descriptions,
                 )
 
         if self.scan_drives: